# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v1.CreateActivity:
    properties:
      attachment:
        type: string
      leadID:
        type: string
      notes:
        type: string
      type:
        type: string
    required:
    - leadID
    - notes
    - type
    type: object
...
  v1.UpdateActivityRequest:
    properties:
      notes:
        type: string
    required:
    - notes
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@crm.com
    name: CRM Support
  description: API untuk sistem CRM (Customer Relationship Management)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CRM Project API
  version: "1.0"
paths:
  /activities:
    post:
      consumes:
      - application/json
      parameters:
      - description: Data aktivitas
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateActivity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat aktivitas baru
      tags:
      - Activities
  /activities/{id}:
    delete:
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus aktivitas
      tags:
      - Activities
    put:
      consumes:
      - application/json
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: Data update aktivitas
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update aktivitas
      tags:
      - Activities
  /activities/{lead_id}:
    get:
      parameters:
      - description: Lead ID
        in: path
        name: lead_id
        required: true
        type: string
      - default: 1
        description: Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List aktivitas by lead
      tags:
      - Activities
...
  /upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File (maks 1MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Upload file attachment
      tags:
      - Activities
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"